+++ src/cashier_frontend/src/components/confirmation-drawer/confirmation-drawer.tsx
@@ -27,7 +27,7 @@ import { ActionModel } from "@/services/types/action.service.types";
 import { ConfirmationPopupLegalSection } from "./confirmation-drawer-legal-section";
 import { isCashierError } from "@/services/errorProcess.service";
 import { useIdentity } from "@nfid/identitykit/react";
-import { useLinkAction } from "@/hooks/link-action-hooks";
+import { useLinkAction } from "@/hooks/useLinkAction";
 import { useProcessAction, useProcessActionAnonymous, useUpdateAction } from "@/hooks/action-hooks";
 import { ConfirmationPopupFeesSection } from "./confirmation-drawer-fees-section";
 import { useIcrc112Execute } from "@/hooks/use-icrc-112-execute";
