+++ src/cashier_frontend/src/hooks/useLinkAction.ts
