+++ src/cashier_frontend/src/pages/edit/[id]/LinkPreview.tsx
@@ -26,7 +26,7 @@ import { getTokenImage } from "@/utils";
 import { Label } from "@/components/ui/label";
 import { useResponsive } from "@/hooks/responsive-hook";
 import { formatNumber } from "@/utils/helpers/currency";
-import { useLinkAction } from "@/hooks/link-action-hooks";
+import { useLinkAction } from "@/hooks/useLinkAction";
 import { useTokens } from "@/hooks/useTokens";
 import {
     ACTION_TYPE,
